v_medium/requirements.txt
kivy
pillow
cryptography
ic-py
pyserial
